+++ bzt/modules/screen.py
@@ -134,10 +134,9 @@ class GUIScreen(BaseScreen):
             pos = 0
             for part in row:
                 txt = part[2]
-                if isinstance(txt, text_type):
-                    strlen = len(txt)
-                else:
-                    strlen = len(txt.decode('utf8'))
+                if not isinstance(txt, text_type):
+                    txt = txt.decode('utf8')
+                strlen = len(txt)
                 self.text.insert(tkinter.END, txt)
                 if part[0] is not None:
                     self.text.tag_add(part[0], "%s.%s" % (idx + 1, pos), "%s.%s" % (idx + 1, pos + strlen))